Create Bacteria_Granule_colocalization_v3.ijm
The code segments and detects bacteria and granule residing within, based on size and shape metrics. It then calculates the overall numbers and sizes of them, and save the results to txt files, the image of overlaid granule and bacteria objects on the raw EM image and associated Fiji roiManager zip-files of the detected granule and bacteria ROIs.